Factor out SearchFilterControl and SelectFilterControl
