+++ src/app/common/components/FilterToolbar/index.ts
@@ -1,2 +1,1 @@
 export * from './FilterToolbar';
-export * from './FilterControl';
